refactor get phone numbers function
