+++ lib/services/twilio.js
@@ -4,9 +4,7 @@ const authToken = process.env.TWILIO_AUTH_TOKEN;
 const client = require('twilio')(accountSid, authToken);
 const formatTime = require('./formatTime');
 
-const Band = require('../models/Band');
-const User = require('../models/User');
-
+const getPhoneNumbers = require('./getPhoneNumbers');
 
 const sendConcertMessage = (concert) => {
   const {
@@ -14,34 +12,20 @@ const sendConcertMessage = (concert) => {
     name,
     startTime,
   } = concert;
-  
-  Band
-    .findById(concert.bandId)
-    .then(async(band) => {
-      await Promise.all(band.members.map(userId => {
-        return User
-          .findById(userId);
-      }))
-        .then(users => {
-          const phoneNumbers = users.map(user => {
-            const phone = user.phone;
-            const part1 = phone.slice(1, 4);
-            const part2 = phone.slice(6, 9);
-            const part3 = phone.slice(10, 14);
-            return `+1${part1}${part2}${part3}`;
+  getPhoneNumbers(concert)
+    .then(phoneNumbers => {
+      return Promise.all(
+        phoneNumbers.map(number => {
+          return client.messages.create({
+            to: number,
+            from: process.env.MY_TWILIO_NUMBER,
+            body: `Band-Tracker\nNEW CONCERT INFO!\n\nConcert: ${name}\nLocation: ${address}\nDate: ${startTime.toDateString()} ${formatTime(startTime)}`
           });
-          Promise.all(
-            phoneNumbers.map(number => {
-              return client.messages.create({
-                to: number,
-                from: process.env.MY_TWILIO_NUMBER,
-                body: `Band-Tracker\nNEW CONCERT INFO!\n\nConcert: ${name}\nLocation: ${address}\nDate: ${startTime.toDateString()} ${formatTime(startTime)}`
-              });
-            })
-          )
-            .then(() => console.log('Messages Sent!'));
-        });
-    });
+        })
+      );
+    })
+    .then(() => console.log('Messages Sent!'));
 };
+
 
 module.exports = sendConcertMessage;
